updated for the database changes
